+++ src/dreamstate/core/queries/unRegisterQueryProvider.test.ts
@@ -1,5 +1,5 @@
 import { QUERY_PROVIDERS_REGISTRY } from "@/dreamstate/core/internals";
-import { queryData } from "@/dreamstate/core/queries/queryData";
+import { queryDataAsync } from "@/dreamstate/core/queries/queryDataAsync";
 import { queryDataSync } from "@/dreamstate/core/queries/queryDataSync";
 import { registerQueryProvider } from "@/dreamstate/core/queries/registerQueryProvider";
 import { unRegisterQueryProvider } from "@/dreamstate/core/queries/unRegisterQueryProvider";