+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 3.4 (*/*/*)
+## 4.0.0 (*/*/*)
   * ContextService removed in favor of ContextManager
   * Provide decorator removed
   * Consume decorator removed
@@ -7,6 +7,8 @@
   * IS_SINGLE removed in favor of scoped storages without global storing // related to HMR problems
   * getCurrent moved to test-utils
   * getCurrentContext moved to test-utils
+  * queryData renamed to queryDataAsync
+  * queryDataAsync stopped array of queries supporting
 
 ## 3.3.2 (4/11/2021)
   * 'partialHotReplacement' parameter for created provider elements that are disabled by default -> allow hot updates and partial reload of context managers